requirement.txt
alabaster==0.7.12
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coverage==6.4.2
crispy-bootstrap5==0.6
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.5
Django==3.2.14
django-allauth==0.51.0
django-ckeditor==6.4.2
django-coverage-plugin==2.0.3
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-extensions==3.2.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-redis==5.2.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
django-taggit==3.0.0
docutils==0.18.1
executing==0.9.0
factory-boy==3.2.1
Faker==13.15.1
filelock==3.7.1
flake8==4.0.1
flake8-isort==4.1.1
fontawesome-free==5.15.4
hiredis==2.0.0
identify==2.5.2
idna==3.3
imagesize==1.4.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.7.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-sugar==0.9.5
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.0.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.3.0
termcolor==1.1.0
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.3.0
types-pytz==2022.1.2
types-PyYAML==6.0.11
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.0
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.1
